order service/requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.0
click==8.3.1
colorama==0.4.6
fastapi==0.124.0
h11==0.16.0
idna==3.11
kafka-python==2.3.0
psycopg2==2.9.11
pydantic==2.12.5
pydantic_core==2.41.5
starlette==0.50.0
typing-inspection==0.4.2
typing_extensions==4.15.0
uuid==1.30
uvicorn==0.38.0
